Move benchmark with specific dataset behind build tag
These benchmarks need some sort of dataset. It doesn't need to be a very
specific one, but 100+ pb.gz files should do it for a start.
